feat: artist module with arquitecture hexagonal (#15)

+++ src/database/migrations/20200807232505-create-person.js
@@ -33,7 +33,7 @@ module.exports = {
           key: 'id'
         },
       },
-      document_number: Sequelize.INTEGER,
+      document_number: Sequelize.BIGINT(20),
       height: Sequelize.DECIMAL(10, 2),
       ip_address: Sequelize.STRING,
       created_at: {
